moved header files next to source files and removed include directory
