--- libhirte/include/common/memory.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#define _cleanup_(x) __attribute__((__cleanup__(x)))
