+++ libhirte/meson.build
@@ -1,25 +1,24 @@
 libhirte_src = [
-    'include/node.h',
-    'include/orchestrator.h',
-    'include/bus/peer-bus.h',
-    'include/bus/user-bus.h',
-    'include/bus/systemd-bus.h',
-    'include/socket.h',
-    'include/common/memory.h',
-    'include/orchestrator/peer-manager.h',
-    'include/service/service.h',
-    'include/service/shutdown.h',
-
+    'src/node.h',
+    'src/node.c',
+    'src/orchestrator.h',
+    'src/orchestrator.c',
+    'src/socket.h',
+    'src/socket.c',
+    'src/bus/peer-bus.h',
+    'src/bus/peer-bus.c',
+    'src/bus/user-bus.h',
+    'src/bus/user-bus.c',
+    'src/bus/systemd-bus.h',
+    'src/bus/systemd-bus.c',
     'src/common/memory.h',
     'src/common/dbus.h',
-    'src/node.c',
-    'src/orchestrator.c',
-    'src/bus/peer-bus.c',
-    'src/bus/user-bus.c',
-    'src/bus/systemd-bus.c',
-    'src/socket.c',
+    'src/common/list.h',
+    'src/orchestrator/peer-manager.h',
     'src/orchestrator/peer-manager.c',
+    'src/service/service.h',
     'src/service/service.c',
+    'src/service/shutdown.h',
     'src/service/shutdown.c',
 ]
