+++ libhirte/src/bus/peer-bus.c
@@ -2,8 +2,8 @@
 #include <errno.h>
 #include <stdio.h>
 
-#include "../../include/bus/peer-bus.h"
 #include "../common/dbus.h"
+#include "peer-bus.h"
 
 char *assemble_tcp_address(const struct sockaddr_in *addr) {
         if (addr == NULL) {
