+++ libhirte/src/bus/user-bus.h
