+++ libhirte/src/common/common.h
