+++ libhirte/src/common/list.h
