+++ libhirte/src/common/memory.h
@@ -3,8 +3,6 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-#include "../../include/common/memory.h"
-
 static inline void *steal_pointer(void *ptr) {
         void **ptrp = (void **) ptr;
         void *p = *ptrp;
@@ -24,5 +22,6 @@ static inline void freep(void *p) {
 
 #define malloc0(n) (calloc(1, (n) ?: 1))
 
+#define _cleanup_(x) __attribute__((__cleanup__(x)))
 #define _cleanup_free_ _cleanup_(freep)
 #define _cleanup_fd_ _cleanup_(closep)
