+++ libhirte/src/service/shutdown.h
